Moved add collection components to modal folder, added refresh boolean when refreshing nft
