+++ src/screens/tabs/NFTs/modals/AddCollection/components/CustomCollection/index.tsx
