+++ src/screens/tabs/NFTs/modals/AddCollection/components/CustomCollection/styles.ts
