+++ src/screens/tabs/NFTs/modals/AddCollection/components/ReviewCollection/index.tsx
