+++ src/screens/tabs/NFTs/modals/AddCollection/components/ReviewCollection/styles.ts
